Add support for inline deserialization
Certain types exist in the system which retain total control over the memory they allocate.  Nearly all container types fit this description, as do certain custom types such as FastImage and certain Eigen types.

In cases where it can be determined that no type in a descriptor heirarchy needs to delegate memory management to a third party, it should be permissible for the user to make use of a simpler syntax that doesn't return an `std::shared_ptr` or require the allocation of a somewhat memory-intensive `Allocation` bookkeeping structure.  This change should also make it trivial to add native support in the future for `std::unique_ptr` and `std::shared_ptr` as embedded members.