Add support for std::unique_ptr
Users that make use of `std::unique_ptr` gain the ability to allocate memory on-demand without needing to create irresponsible descriptors.  The memory pointed to by `std::unique_ptr` obeys all of the same rules for aliasing, even if there are dumb pointers elsewhere in the system that refer to this `std::unique_ptr`, and even if those pointers are encountered first during serialization or deserialization.