Implement protobuf output formatter
This formatter will allow LeapSerial structures to target the protobuf wire format.  Protobuf wire-level compatibility makes it possible to interop with protobuf comms layers without needing to generate a .proto file.

This formatter is only meant to be an initial implementation, and will be tested but not to the extreme.  Subsequent pull requests should ensure the implementation is adequately tested for general use.