Fix extremely inefficient use of GetDescriptor
Vector and map serializers are constructing temporary `field_serializer_t` items on the stack every time they are called for.  This practice is extremely inefficient and is one of the reasons that the `field_serializer_t::GetDescriptor()` method was provided to begin with.  Fix these routines so they call `GetDescriptor()` as the other serializers do.