Provide more useful static assertions when performing descriptor instantiation
We should probably allow pattern matching in all `primitive_serial_traits` types and simply change the inheritance to indicate whether the matched type is valid or not.  We only throw a hard error if a user attempts to instantiate one of the static methods in the type, which would imply an attempt to use rather than a simple query on the availability of a serializer.