Adapter streams should take ownership of their underlying fields
This is necessary to allow a single stream to completely wrap the interior stream.  Otherwise, we wind up with memory management issues everywhere.  Furthermore, there should really be a concept of exclusive ownership, here--if someone attempted to directly access the underlying stream, we could possibly wind up with a state inconsistency that would be tough to resolve in the general case.