Fix vector-of-optionals defect
Currently the LeapSerial archiver does not write size values down for entries of an array or a map.  This means that archivers which rely on exterior knowledge of the size--such as the descriptor archiver--do not know when the object is "finished", and therefore can only deserialize those parts of the object which are non-optional.

Correct this while preserving backwards-compatibility by setting the high bit of the element count field on the vector when serializing mutable types.  During deserialization, if this bit is set, a size integer is extracted from the stream before every deserialized array entry.  If the bit is cleared, deserialization proceeds as normal.  Also add a concept of "immutable" to identify those types for which a size field is redundant.