Add benchmark project
Reports of slow performance in the AES encryption stream cipher have made it necessary to add a project to benchmark its performance and ensure that it approaches the performance that could be achieved with a plaintext read.  IO should be the limiting factor with a block encryption system such as AES, not compute.