Fix inconsistent stream EOF behaviors
Some streams are throwing exceptions when the EOF is reached, others set the EOF bit exactly when the last byte is read from the stream, and others still set the EOF bit when an attempt is made to read beyond the end of the file.

Standardize on the behavior defined by the STL:  The EOF bit indicates whether the last read operation read past the end of the file.  This bit gets set even if the last read operation successfully moved some data.